SoniqueBackend/requirements.txt
flask
pymongo
flask_cors
tinydb
fitz
google-generativeai
python-dotenv
